requirements.txt
numpy==1.16.4
Pillow==6.2.1
